Fix docker recommendations
