+++ src/main/java/com/kruize/analysis/AnalysisImpl.java
@@ -18,11 +18,14 @@ package com.kruize.analysis;
 
 import com.kruize.metrics.AbstractMetrics;
 import com.kruize.metrics.MetricCollector;
+import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 
+import java.math.RoundingMode;
 import java.text.DecimalFormat;
 import java.util.ArrayList;
 
-public abstract class AnalysisImpl<T extends AbstractMetrics> implements Analysis<T> {
+public abstract class AnalysisImpl<T extends AbstractMetrics> implements Analysis<T>
+{
     public void calculateCpuLimit(T instance)
     {
         double maxCpu = 0;
@@ -47,6 +50,84 @@ public abstract class AnalysisImpl<T extends AbstractMetrics> implements Analysi
         instance.setCurrentCpuLimit(cpuLimit);
     }
 
+    public void calculateCpuRequests(T instance)
+    {
+        final int PERCENTILE = 80;
+        final int INDEX = MetricCollector.CPU_INDEX;
+
+        ArrayList<MetricCollector> metrics = instance.metricCollector;
+        ArrayList<MetricCollector> percentileList = new ArrayList<>();
+
+        if (metrics.size() == 0) {
+            instance.setCurrentCpuRequests(-1);
+            return;
+        }
+
+        DescriptiveStatistics referenceValues = new DescriptiveStatistics();
+
+        for (MetricCollector metric : metrics) {
+            double value = metric.getFromIndex(INDEX);
+            referenceValues.addValue(value);
+            System.out.print(value + "\t");
+        }
+
+        double percentileValue = Statistics.getPercentile(referenceValues, PERCENTILE);
+        System.out.println(PERCENTILE + "th percentile is " + percentileValue);
+
+        for (MetricCollector metric : metrics) {
+            if (metric.getFromIndex(INDEX) >= percentileValue) {
+                MetricCollector temp = MetricCollector.Copy(metric);
+                DecimalFormat singleDecimalPlace = new DecimalFormat("#.#");
+                singleDecimalPlace.setRoundingMode(RoundingMode.CEILING);
+                double targetValue = temp.getFromIndex(INDEX);
+                temp.setForIndex((Double.parseDouble(singleDecimalPlace.format(targetValue))), INDEX);
+
+                percentileList.add(temp);
+                System.out.println(temp.toString());
+            }
+        }
+
+        double cpuRequests = Statistics.getMode(percentileList, INDEX);
+        instance.setCurrentCpuRequests(cpuRequests);
+    }
+
+    public void calculateMemRequests(T instance, int referenceIndex, int targetIndex)
+    {
+        final int PERCENTILE = 80;
+        final int ROUND_TO_MUL_OF = 5;
+
+        ArrayList<MetricCollector> metrics = instance.metricCollector;
+        ArrayList<MetricCollector> percentileList = new ArrayList<>();
+
+        if (metrics.size() == 0) {
+            instance.setCurrentRssRequests(-1);
+            return;
+        }
+
+        DescriptiveStatistics referenceValues = new DescriptiveStatistics();
+
+        for (MetricCollector metric : metrics) {
+            double value = metric.getFromIndex(referenceIndex);
+            referenceValues.addValue(value);
+            System.out.print(value + "\t");
+        }
+
+        double percentileValue = Statistics.getPercentile(referenceValues, PERCENTILE);
+        System.out.println(PERCENTILE + "th percentile is " + percentileValue);
+
+        for (MetricCollector metric : metrics) {
+            if (metric.getFromIndex(referenceIndex) >= percentileValue) {
+                MetricCollector temp = MetricCollector.Copy(metric);
+                temp.setForIndex(roundToNearestMultiple(temp.getFromIndex(targetIndex), ROUND_TO_MUL_OF), targetIndex);
+                percentileList.add(temp);
+                System.out.println(temp.toString());
+            }
+        }
+
+        double memRequests = Statistics.getMode(percentileList, targetIndex);
+        instance.setCurrentRssRequests(memRequests);
+    }
+
     public void calculateMemLimit(T instance)
     {
         double spike;
@@ -61,7 +142,7 @@ public abstract class AnalysisImpl<T extends AbstractMetrics> implements Analysi
         }
 
         for (MetricCollector metricCollector : metrics) {
-            double mem = metricCollector.getFromIndex(MetricCollector.CPU_INDEX);
+            double mem = metricCollector.getFromIndex(MetricCollector.RSS_INDEX);
             rssValues.add(mem);
             if (maxMem < mem)
                 maxMem = mem;
@@ -103,4 +184,13 @@ public abstract class AnalysisImpl<T extends AbstractMetrics> implements Analysi
         instance.setCpuLimit(Math.max(instance.getCpuLimit(), currentCpuLimit));
         instance.setCpuRequests(Math.max(instance.getCpuRequests(), currentCpuRequests));
     }
+
+    // function to round the number to multiple of number specified
+    @SuppressWarnings("SameParameterValue")
+    private static double roundToNearestMultiple(double number, double multipleOf)
+    {
+        return multipleOf * (Math.ceil(Math.abs(number/multipleOf)));
+    }
+
+
 }
